use package testenv and rename methods
