+++ components/serverless/internal/controllers/kubernetes/namespace_controller_test.go
@@ -22,8 +22,8 @@ import (
 func TestNamespaceReconciler_Reconcile(t *testing.T) {
 	//GIVEN
 	g := gomega.NewGomegaWithT(t)
-	k8sClient, testEnv := testenv_test.SetUpTestEnv(t)
-	defer testenv_test.TearDownTestEnv(t, testEnv)
+	k8sClient, testEnv := testenv.Start(t)
+	defer testenv.Stop(t, testEnv)
 	resourceClient := resource.New(k8sClient, scheme.Scheme)
 	testCfg := setUpControllerConfig(g)
